+++ resources/views/computer/create.blade.php
@@ -17,9 +17,9 @@
         </x-create.abschnitt>
 
         <x-create.abschnitt :titel="__('Fernwartung')">
-            <x-create.singlerow :label="__('Rustdesk ID')" name="remoteID" />
+            <x-create.singlerow :label="\App\Models\Setting::fernwartung()['id_label']" name="remoteID" />
 
-            <x-create.singlerow :label="__('Rustdesk Passwort')" name="remotePassword" />
+            <x-create.singlerow :label="\App\Models\Setting::fernwartung()['password_label']" name="remotePassword" />
         </x-create.abschnitt>
 
         {{-- Weitere IP-Adressen und Zugangsdaten haengen am gespeicherten Objekt;