+++ resources/views/vm/edit.blade.php
@@ -21,9 +21,9 @@
         </x-create.abschnitt>
 
         <x-create.abschnitt :titel="__('Fernwartung')">
-            <x-create.singlerow :label="__('Rustdesk ID')" name="remoteID" :default="$vm->remoteID" />
+            <x-create.singlerow :label="\App\Models\Setting::fernwartung()['id_label']" name="remoteID" :default="$vm->remoteID" />
 
-            <x-create.singlerow :label="__('Rustdesk Passwort')" name="remotePassword" :default="$vm->remotePassword" />
+            <x-create.singlerow :label="\App\Models\Setting::fernwartung()['password_label']" name="remotePassword" :default="$vm->remotePassword" />
         </x-create.abschnitt>
 
         <x-create.abschnitt :titel="__('Dienste')" :hinweis="__('Aus dem Katalog wählen oder eigene ergänzen')">